Refactor Item constructors.
Gave theCas a default value of zero since this is what the constructors
did already. Got rid of the initialize method because it obfuscated what
the constructors were doing and allowed the use of different variants of
string:assign.

Added setData method and moved the only complex logic from initialize
into that.

Change-Id: Idea3bdcb54944b889b035a0c3f85a3ee03f2c3d7
Reviewed-on: https://review.northscale.com:8443/1328
Reviewed-by: Dustin Sallings <dustin@spy.net>
Tested-by: Dustin Sallings <dustin@spy.net>