Achieved still yet deeper understanding of CRLF handling in memcached.
  memcached asks for two bytes more than it needs and adds its own CRLF.

  The engine should store the things it needs only.

  *EXCEPT*

  incr has to always add CRLF to the end of the value it stores.

The consequence of this is that we will persist two bytes more than
was strictly required for every record.  The value is opaque to us, so
we're just going to blindly store what memcached told us because *it*
deals with the complexity around the additional CRLF (unless it's an
incr or decr).

Change-Id: I3c6618a596fdd0721589f7a413c5ffc2684a2986
Reviewed-on: http://review.northscale.com:8080/274
Reviewed-by: Sean Lynch <seanl@literati.org>
Tested-by: Eric Lambert <eric.d.lambert@gmail.com>
Tested-by: Dustin Sallings <dustin@spy.net>