Support for resuming connections blocked on vbucket state.
Connections that are blocked when a state goes from pending -> active
are awaked using notify_io_complete and are allowed to finish their
jobs.

Note that this doesn't come with an engine test because our engine
testing framework isn't quite sophisticated enough to model this
currently (or if it is, I can't quite figure it out).

I tested this by putting vbucket 3 into a pending state and then
simultaneously starting 10 operations attempting to set distinct
values into it and having another connection attempting to read a
value from it.  They all blocked.

Then I put vbucket 3 into active state and watched all the events
fire.  The get actually failed immediately (yay concurrecy).  This is
not particularly bad as it's a sort of undefined state, but it worked
immediately afterwards (once the commands all completed processing).

Change-Id: Ic245ad9e70e3dd4a9b49de9a7a52ade32db81922
Reviewed-on: http://review.northscale.com:8080/405
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Dustin Sallings <dustin@spy.net>