Enforce max memory usage - bug 1204
This uses a new parameter max_size (in bytes -- defaulting to
something large) to be a rough memory ceiling.

It refuses to store items into the hash table if the
sizeof(StoredValue) + key.length() + val.length() > this maximum
amount.

It does not refuse on item resizes, incr resizes or similar types of
ops, onnly complete new allocations.

It also does not take into account other overhead such as queues, the
hashtable sizes themselves, and probably a bit of additional overhead
in the item storage (such as the size of the shared ptr).

This change also introduces a new stat for viewing the current size of
the allocations it's using to make its decisions (made it slightly
easier to play with).

Unit test sets a small size and verifies that it can add an item, fail
to add a second (with ENGINE_ENOMEM) when it would go over the limit,
and then can succeed to add that second item after deleting the first.

Change-Id: I36e96fc75ead09dcdc69047046fdc541d62d594c
Reviewed-on: http://review.northscale.com:8080/581
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Dustin Sallings <dustin@spy.net>