Backport: Reconnect tap connections that stay idle too long
If a tap connection doesn't receive a message for a configurable delay
(specified by tap_idle_timeout (default 600secs)), the tap client
connection close and tries to reconnect to the server)

This backport includes a new mechanism to send a NOOP from the tap
walker.

This is part of the fix for bug 1357.

This change has been backported from
8b01dbcb93577fa3f4194bb32c8f0365f8d0f1e8
http://review.northscale.com:8080/#change,573

The cherry-pick had the following conflicts:

	ep_engine.cc
	ep_engine.h
	tapclient.cc

Change-Id: I1d7074e381064218f1fb7e8409cd6ec8173dbff2
Reviewed-on: http://review.northscale.com:8080/588
Reviewed-by: Dustin Sallings <dustin@spy.net>
Tested-by: Dustin Sallings <dustin@spy.net>