Bug 1428 - Ensure the vbucketmigrator finishes quickly in all cases
There was a race condition where the backfill thread, when there was
nothing in the source list would notify the io thread to go do some
work.

There are, however, two points where it might be waiting.

If the tap IO thread is waiting in the second one, it'll notify all
the paused tap connections and that will cause them to proceed.

If you notify the first one, it will do some cleanup, look around and
see that all connections have empty queues and are in the paused state
and decide it doesn't have anything to do for a while (until the next
time a record comes in to stimulate the tap queue or it's time to
transmit a keepalive NOOP.

This change resets the NOOP time so there will immediately be a NOOP
available for any tap connection that does not have data pending after
each backfill finishes.  This is a fairly rare situation and, in its
worse case, ends up with an extra packet transmitted on idle
connections.

Change-Id: Ia766651418b261ad2de8a664415285a3b83b509b
Reviewed-on: http://review.northscale.com:8080/652
Tested-by: Sean Lynch <seanl@literati.org>
Reviewed-by: Sean Lynch <seanl@literati.org>