Prevent concurrent backfills for a single tap client - bug 1635
Hash table visitors can now periodically ask if they should keep going
the backfill visitor can use state tracked within the server to know
that a backfill is still applying to the connection it started with.

This check could fail if a connection is recycled for the exact same
named tap stream -- in that case, it will degrade to duplication
again, taking an already rare occurrence and make it even more rare.

Change-Id: I16bda3331dc767c1525f25cf8b196f62c339676c
Reviewed-on: http://review.northscale.com/965
Tested-by: Matt Ingenthron <matt@northscale.com>
Reviewed-by: Matt Ingenthron <matt@northscale.com>