Disable dedup support for backfilled tap data - bug 1654
When incrementally filling, the cost shifted *tremendously* towards
maintaining dedup data and moving queues in and out.

While it's possible that a backfill under load will introduce
duplicate items to the tap queue, this will only be duplicated with
what the backfiller is doing, so it won't be more than one extra.
This is well worth the cost savings.

Change-Id: Id7da265db72691f8d2c81fac5be731b0a1dbb20d
Reviewed-on: http://review.northscale.com/986
Reviewed-by: Matt Ingenthron <matt@northscale.com>
Tested-by: Matt Ingenthron <matt@northscale.com>