Throttle backfills while consumers are slow - bug 1635
This is very coarse throttling, but it does prevent the backfill from
running to completion when clients are too slow to keep up, saving
precious memory and freeing up locks and CPU for things that are
needed.

Enough of a backlog is generally maintained that nothing should be
waiting for this to complete.

Change-Id: Id431beeefad6f20cc1ef6ed6b70823102be0fd97
Reviewed-on: http://review.northscale.com/989
Reviewed-by: Matt Ingenthron <matt@northscale.com>
Tested-by: Matt Ingenthron <matt@northscale.com>