A safer io completion signaler.
A copy of the tapConnectionMap was being iterated without a lock to
figure out which connections needed to be signaled.

That in itself is thread safe, but the iteration was checking
properties on one of the tap connections in an unsafe way.  We now do
that filtering within the lock as well.

It is still possible that the signaling of an unhandled cookie could
lead to issues, but it's very unlikely and the server doesn't provide
us an API to avoid a race, so the burden is on it to prevent
notify_io_complete calls with invalid cookies.

Change-Id: Iac694cbe8ba4ea548e21ba35b2a3e3a18773c646
Reviewed-on: http://review.northscale.com/996
Reviewed-by: Matt Ingenthron <matt@northscale.com>
Tested-by: Matt Ingenthron <matt@northscale.com>