Verify we have a tap connection to remove at hangup - bug 1655
I was able to demonstrate this bug by connecting and disconnecting
once per second and suspending the process through a few cycles.  I
will have a mapping for the cookie, but no tap client associated with
it.

The problem was that map[x] has a side-effect when x is not in the
map.  It creates it.  That was leaving a null tap connection for a
connection that didn't quite exist yet and would never exist because
it was disconnecting.

The disconnect code guards against this and the walker prevents it
from happening.

Change-Id: I92873e33ffc51ca00e8635cf02c021d032078a62
Reviewed-on: http://review.northscale.com/1003
Reviewed-by: Matt Ingenthron <matt@northscale.com>
Tested-by: Matt Ingenthron <matt@northscale.com>