Mostly removed getFromUnderlying.
This is used internally by "stats vkey" which currently doesn't supply
a vbucket ID, so it's not very useful in its current state.  It'd
require some work to get it updated, so I ejected it until we can get
a better form for it and get some testing around it.

Change-Id: I4151d0f138e0e6a4c757eb4e968ac9a4d32a1e41
Reviewed-on: http://review.northscale.com/1052
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Trond Norbye <trond.norbye@gmail.com>