Improved the logic for tap connection cleanup
Instead of nuking the tap connection object if another connection
tries to connect to an existing tap connection, we create a temporary
queue and bind the old connection to that queue and notifies the old
cookie. This cause the core to try to get the next item on that tap
stream, but instead of returning new events it tells memcached to
disconnect the tap stream.  This allows memcached to clean up all
resources allocated for the connection.

Change-Id: I288241e48c281dc6167cf5bfa2fd8e1a68a60e55
Reviewed-on: http://review.northscale.com/1172
Reviewed-by: Matt Ingenthron <matt@northscale.com>
Tested-by: Matt Ingenthron <matt@northscale.com>