Improved the logic for tap connection cleanup
Instead of nuking the tap connection object if another connection
tries to connect to an existing tap connection, we create a temporary
queue and bind the old connection to that queue and notifies the old
cookie. This cause the core to try to get the next item on that tap
stream, but instead of returning new events it tells memcached to
disconnect the tap stream.  This allows memcached to clean up all
resources allocated for the connection.

Change-Id: Ifee14a39e64396341b08ad152d061e64dd939feb
Reviewed-on: http://review.northscale.com/1224
Reviewed-by: Sean Lynch <seanl@literati.org>
Tested-by: Sean Lynch <seanl@literati.org>