More configurability around automatic value ejection.
When the memory usage exceeds a high water mark, we'll randomly throw
away some items in such a way that should get us approximately to a
low water mark.

The high and low water marks are configurable by mem_low_wat and
mem_high_wat respectively.

If not configured, the low water mark is 60% of the max amount of
memory and the high water mark is 75%.

Change-Id: Ieb68901a58e5e6a60f61bf7ca31492ce76269540
Reviewed-on: http://review.northscale.com/1180
Reviewed-by: Sean Lynch <seanl@literati.org>
Tested-by: Sean Lynch <seanl@literati.org>