Disable extra mutex check for win32
The pthread implementation we use on win32 doesn't support the
construct: pthread_t id = pthread_self();

The code is just an extra error check added at a time we had weird
mutex problems, and not really necessary for the code to work. If we
have mutex problems on win32 we should figure out how to implement the
same functionality on win32.

Change-Id: I6035b9e5ddf1441dc25f4ffd61d10e47b6bbf4c0
Reviewed-on: http://review.northscale.com/1279
Reviewed-by: Dustin Sallings <dustin@spy.net>
Tested-by: Dustin Sallings <dustin@spy.net>