Shut down connections without data properly
There was a race condition on how to shut down a tap connection
it there wasn't any data to be transmitted on the tap queue.

The hack to "modify" the nextTapNoop timer didn't completely solve
this problem. There is no point of sending a ton of noops, so we
should rather send them at a fixed rate with a regular interval. We're
sending 3 noops per timeout interval, so all clients should get them
before deciding to disconnect anyway.

Change-Id: Ie5b0f7b53bf4a03a298d33fc01903dec22c101a6
Reviewed-on: http://review.northscale.com/1401
Tested-by: Dustin Sallings <dustin@spy.net>
Reviewed-by: Dustin Sallings <dustin@spy.net>