bug 1699 - curr_items should only include items in active vbuckets
This is accomplished by removing the curr_items stat counter and
replacing it with an atomic hashtable item size, then calculating the
curr_items stat by visiting all the active vbuckets at request time.

curr_items stat test updated to show disabling and reenabling a
vbucket's affecting the count.  hash_table_test was updated to show
that all areas where we verify the size of the hashtable (every where
we mutate it) keep the size counter in sync.

Change-Id: Ib173ad06590628545771bf497cfa85761a558593
Reviewed-on: http://review.northscale.com/2314
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>