Refactor due to upstream interface changes
The cookie-based access functions is moved to it's own structure in
the server interface, and this commit reflect that change.

A lot of the places also passed in the server api as a parameter to
the function, but it is also availabe through a member in the class,
so in those cases I just dropped the parameter.

Change-Id: I094ece8890e898e953d32ad04a397daf5d260175
Reviewed-on: http://review.northscale.com/2424
Reviewed-by: Dustin Sallings <dustin@spy.net>
Tested-by: Dustin Sallings <dustin@spy.net>