Bug 2360 - Do not persist expired objects
If an object is expired (or will be soon), we should avoid persisting
it.  This provides a new stat (ep_item_flush_expired) that
represents the number of times an item is not flushed due to the
expiry of the item. This also introduces a new config param
(expiry_window) which is used for checking if an item's expiry time
is less than current_time + expiry_window.
The new test case (test_expiry_flush) has been added for testing
this enhancement.

Change-Id: I6aed62edc30a5c2378dc66f8240f74d356ef4bf6
Reviewed-on: http://review.northscale.com/2441
Reviewed-by: Dustin Sallings <dustin@spy.net>
Tested-by: Dustin Sallings <dustin@spy.net>