Bug 2256 - stats vkey causes assert
stats vkey is now supported. The protocol is adapted to allow the client
to query vkey stats through key + vbid. For example, "vkey k1 0" means
stats for key "k1" in vbucket "0". If vbid is missing, then we assume
that vbid is "0". In this way, we can support the backward
compatibility. The protocol for stats key is adapted in the same way.

Change-Id: I3b80abfd449a3a89f40efe3c0f98365b36eb7e43
Reviewed-on: http://review.northscale.com/2549
Reviewed-by: Dustin Sallings <dustin@spy.net>
Tested-by: Dustin Sallings <dustin@spy.net>