Bug 2418 - sporadic test failure: test_flush_stats
I found that when a new item is inserted, "ep_overhead" stat remains
the same if the flusher picks that item to store it into the disk
before we get the stats again. This is is a normal behavior and can
happen if there are no tasks in the dispatcher queue, which have a
higher priority than the flusher. I fixed this issue in the unit
test module by checking if "mem_used" stat is equal to the sum of
"ep_overhead" and "ep_total_cache_size" after inserting a new item.

Change-Id: Ie9d91ddc3db8b9cb270c3f7fec809f69cb1c01e6
Reviewed-on: http://review.northscale.com/2583
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>