bug 2433 - Option to continue when we can't load all data
ep_warmup_oom increments whenever we encounter an OOM-based load
failure.  If failpartialwarmup is true at startup, we will fail to
load all data if it would require us to exceed our required memory,
but we will have noted the condition so a management system can decide
what to do about it.

Change-Id: I71a5eb32076327b57f9f851e55a2b29ddc80070c
Reviewed-on: http://review.northscale.com/2591
Tested-by: Dustin Sallings <dustin@spy.net>
Reviewed-by: Sean Lynch <seanl@literati.org>