Track StoredValue size() with alignment.  Partial bug 2332.
The size() function on StoredValue did not take into account any
alignment overhead requirements on the given architecture.  With
this change it does.

This fix (via Trond) helps with bug 2332, but it does not completely
solve it.  The improvement is correct and welcome, so I'm submitting
it now as a partial change.

Change-Id: I6b76d75e3909deae9bafa703962b1eafa2cd3bf5
Reviewed-on: http://review.northscale.com/2608
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Reviewed-by: Dustin Sallings <dustin@spy.net>
Tested-by: Dustin Sallings <dustin@spy.net>