bug 2509 - clear deleted items from the ht in deletion cb
If there's an error of some sort in the deletion itself, it's possible
to end up with a time when an item exists on disk and is still
resident in memory.

With this change, we keep the item in the hash table until we have
positive confirmation that it has been deleted from disk (and, of
course, that it's still deleted in memory).

Change-Id: I0a264121b764d19408ea6cdddf370b3f25843cd1
Reviewed-on: http://review.northscale.com/2740
Reviewed-by: Matt Ingenthron <matt@northscale.com>
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Dustin Sallings <dustin@spy.net>