Bug 2524 - Global tap mutex comes out of hot lock contention
Every operation that required some sort of data mutation would end up
calling addEvent that grabbed the tapNotifySync. This mutex is held by
the tap notify loop that does more stuff with the mutex held, causing
a potential long wait time on this mutex. THis patch changes the logic
to use a temporary list where we put these events in and let the tap
notify thread populate them while it already got the lock.

Change-Id: I076b6675ce4d4855eddab9fd91df3e2bac85dfd1
Reviewed-on: http://review.northscale.com/2840
Reviewed-by: Dustin Sallings <dustin@spy.net>
Tested-by: Dustin Sallings <dustin@spy.net>