bug 2623 - Handle deletion events for items that aren't on disk
The deletion callback was going to disk with a rowid of -1 which would
never result in a successful delete and would cause the object to be
stuck in memory since there's no positive confirmation that it's safe
to be removed.

Also, since it's not going to do anything, we'll bypass the deletion
when we know there's no rowid and just call the callback directly so
the item can be physically deleted if it's still logically deleted
when we get there.

Change-Id: If21ae1795b74d1f5cb5d7c342f11ef0448d48521
Reviewed-on: http://review.northscale.com/3054
Tested-by: Dustin Sallings <dustin@spy.net>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>