bug 2656 - fix prioritization of future scheduled events
This splits the task priority queue into two -- one for "ready" tasks
to be executed in priority order and another for "future" tasks which
are to be added to the "ready" queue as soon as their due date
permits.

This change will increase our gettimeofday calls around scheduling
since every task should now have a "due" date, but this is only at
creation, rescheduling, and execution times.

Change-Id: I600641ebd5b4764e140d4bde5bc9e243f7f99aba
Reviewed-on: http://review.northscale.com/3156
Reviewed-by: Sean Lynch <seanl@literati.org>
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Tested-by: Sean Lynch <seanl@literati.org>