Bug 2865: TAP replication is unbalanced with two clients
There was no kind of flow control for access to the dispatcher,
causing one of the threads to be able to fill it up with all of the
items currently available from the item queue (I've seen more than
50000 requsts there). This patch adds a (default 500) max limit to
the number of pending requests to the dispatcher before it must
wait for any of them to become available. This ensures that both
tap connections will get at least _some_ data. A better "load
balancing" is needed..

To use another max limit add "tap_bg_max_pending=nnn" to the
engine configuration.

ex:
/opt/membase/bin/memcached -E ep.so \
    -e dbname=/usr/tmp/membase.sqlite\;cache_size=1024M\;tap_bg_max_pending=100

Change-Id: Ic9adcffd298e6063aecf3d4669e9aca4a6b8fe81