Item pager that removes items with the old vbucket version from disk
If a membase server is crashed in the middle of vbucket deletion and
creation, there is a high chance of having items with the old
vbucket version in disk. This feature incrementally removes all of
those old items with the lowest priority to minimize the impact on
other bg jobs with higher priorities.

Change-Id: Iad4530fa6f1f64d8730761869211af0b5f229b09
Reviewed-on: http://review.membase.org/3444
Reviewed-by: Dustin Sallings <dustin@spy.net>
Tested-by: Dustin Sallings <dustin@spy.net>