Adjusting a chunk range size dynamically during vbucket deletion
This adjusts a chunk range size dynamically based on
(1) time spent by deleting the current chunk and (2) threshold
execution time (engine parameter) on deleting a single chunk. For
example, assume that (1) it takes 250 msec to delete the current
chunk with the range size 2000  and (2) the threshold execution
time is set to 1 sec. Then, the range size of the next chunk
will be increased to 2000 * (1 sec / 250 msec) = 8000.
If the time spent by deleting this new chunk is 2 sec, then the
range size of the next chunk will be decreased to
8000 * (1 sec / 2 sec) = 4000.

Change-Id: Iafb0b84ec52ee12f49a636d2ec453348a3b35e17
Reviewed-on: http://review.membase.org/3492
Reviewed-by: Dustin Sallings <dustin@spy.net>
Tested-by: Dustin Sallings <dustin@spy.net>