Refactor: replace tap_idle_timeout with tap_noop_interval
tap_idle_timeout used to be the number of seconds the internal tap
client would be idle before it would give up and reconnect. We would
send 3 noops in that interval to "ping" the client to keep if from
timing out.

The internal tap client is replaced by vbucketmigrator, so it doesn't
make sense to keep the configuration variable anymore with it's
current name. Describing it in the document would only confuse users:

"Set tap_idle_timeout to 3 times the number of seconds that you want
ep_engine to send a noop to the client to keep it from timing out"

This patch renames the config parameters to tap_noop_interval, but
preserves backwards compatibility by accepting tap_idle_timeout as
"tap_noop_interval * 3"

Change-Id: I38793ff49a3a6c4a342a35ef206b765b2f44be54
Reviewed-on: http://review.membase.org/3520
Reviewed-by: Dustin Sallings <dustin@spy.net>
Tested-by: Dustin Sallings <dustin@spy.net>