Misc fixes for getl command support
1. arithmetic operations for a locked key were resulting in a crash
because store op always returned EXISTS. This caused the arithmetic
op to be invoked recursively endlessly resulting in a stack overflow.
The correct fix for this should be to perform a getlocked op before
attempting incr/decr on the data. For now return with error if the
object is locked

2. A delete operation on an locked object needs to be disallowed,
failure to do so results in a crash. memcached wrapper should
be able to handle TMPFAIL, currently it can only return NOT_FOUND
or DELETED.

3. flags value needs to be converted to host byte order before
passing the value to the ascii response handler

4. move the check for the object being locked before determining
whether the object is resident in EventuallyPersistentStore::getLocked

- Added testcases for 1 & 2 to test_getl

Change-Id: Ia7f4b83e1feb3fa9ef5d3a01f8932f07fe2829d1
Reviewed-on: http://review.membase.org/3521
Reviewed-by: Dustin Sallings <dustin@spy.net>
Tested-by: Dustin Sallings <dustin@spy.net>