MB-2852:  Let vb snapshot run before trying to lock out the DB.
In the real world, the DB would be locked normally, and then unlocked,
and then things would be OK.  In this test, we're waiting for the
thing to go and the two lock each other out.

MB-2663 makes it easier to detect when this happens so we can target
our test more accurately, but in the meantime, I'm just waiting long
enough for that initial task to have run.

I also put an alarm on the child so that if this happens again, it'll
at least clear itself up sensibly.

Change-Id: Ib76b1237544fca6aa2c23e04d19f0989fb23cbaf
Reviewed-on: http://review.membase.org/3694
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>