Cleaned up dispatcher notifications and idle task interaction.
A race condition allowed an idle task to start as another task was
being added in such a way that the idle task wasn't notified and the
job execution was delayed.

This mostly showed up in slow execution of tests (the generated tests
were the most likely since most of them were here), but would happen
any time an idle task and another task interacted.

Change-Id: I06fd0ff8370acde2dac613f9168c14c532dfdd26
Reviewed-on: http://review.membase.org/3863
Reviewed-by: Toby DiPasquale <toby@cbcg.net>
Tested-by: Trond Norbye <trond.norbye@gmail.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>