Set default threading mode back to 2.
This was required to be 1 briefly as there was concurrent access to
the same connection and prepared statements.

Now, concurrent access is done via multiple sets of connections and
statements, so even though we'll be accessing the same database
potentially, we'll be doing it without common structures.

This is still runtime configurable.

Change-Id: I3918e00fb4f44a8a3369e715d146576c4eb6f6db
Reviewed-on: http://review.membase.org/3865
Tested-by: Trond Norbye <trond.norbye@gmail.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>