Must destroy strategy before database.
This was showing up with malloc scribble on OS X.  If we destroy the
database *after* the strategy, the database shutdown procedure tries
to make use of the strategy that has already been deallocated.

Clearly there are some leaky abstractions here, but we've got test
tools that are at least pointing out when we make little mistakes like
this.

Change-Id: Ia88af5faaed6a30ed6b332f9f627217ea54bc6cf
Reviewed-on: http://review.membase.org/4028
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>