Use :memory: when validating engine handle.
This makes it slightly faster, but more importantly it means that
we'll clean up any leftover databases before running the remaining
tests.

Occasionally, I'll leave a giant db in the test location and try to
run a test and find that it spends forever trying to run the first
test since the initial hash table is tuned to be too small to load all
the data.  We *could* run the cleanup as a test setup, but that'd
involve doing too much stuff.

Change-Id: I6d5bbdef98c2f417ac17a91c7f7a7a04dd292e79
Reviewed-on: http://review.membase.org/4088
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Dustin Sallings <dustin@spy.net>