MB-3154 - Fix to DB update failures caused by the DB reset operation.
After the DB reset operation is executed, there are lots of database
update failures because we sort all the persistence tasks including
the DB reset operation by their shard ids and row ids.

This resolves the above issue by clearing the persistence queue and
pushing the DB reset operation into the front of the persistence
queue.

Change-Id: Iff4d7cbbf4f0e4cc0e526fc5376118bb74c87a90
Reviewed-on: http://review.membase.org/4197
Tested-by: Dustin Sallings <dustin@spy.net>
Reviewed-by: Dustin Sallings <dustin@spy.net>