Fix to the data loss in supporting tap disk backfill per vbucket.
We observed that there are data loss issues in replicas because
a tap background fetch job is skipped when the tap disk backfill
operation does not even run. This change fixes the issue by
scheduling a tap background fetch job iff the tap disk backfill
per vbucket is not supported or not running now.

Change-Id: I66a5c9b31d9c1c2aca2c4b89f796814d1ffea6af
Reviewed-on: http://review.membase.org/4213
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Reviewed-by: Dustin Sallings <dustin@spy.net>