MB-3286, MB-3252: unconditionally declare expired objects dirty
An expired object would previously be returned as NOT_FOUND and would
therefore not be queued dirty and not actually cleaned up from disk.

The object would not be available, but would continue to consume disk
space *and* hash table space since nothing could pull it from the
logical deleted state within the hash table.

Objects that are expired on disk will still be loaded (MB-3286 is
really asking describing a number of different things), and the count
won't immediately go down without the objects being requested or until
the expirating paging timer run.

Change-Id: Ifa4b0f79579cc0cc6b4e7d8ab746b7186372b710
Reviewed-on: http://review.membase.org/4228
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>