MB-3269 Replica is very slow after fail over and rebalance
This was caused by using the wrong variable to determine when to
request an ACK from the other end. The variable we used would be
incremented every time we tried to fetch an item from our message
queue, but the item would not be sent on the wire if the item wasn't
in memory. This resulted in that we only checked a subset of
the values if it was time to insert an ack...

The test to see if the ack window was full checked the
distance between the current sequence number and the last sequence
number we received.

Change-Id: Iaa2020cdceb34f24f164919e942704073923909f
Reviewed-on: http://review.membase.org/4241
Tested-by: Dustin Sallings <dustin@spy.net>
Reviewed-by: Dustin Sallings <dustin@spy.net>