MB-3296: Delete any expired objects at startup.
We will load all items we find on disk, but expired items will be
unavailable.  This would normally take care of itself, but the system
would consume more resources than necessary for up to an hour.

This change detects that a loaded item is expired and immediately
deletes it as if a client had requested it.

Change-Id: Ibe0748d8fc57903ce57dcb2b96f0093d4f98ccfa
Reviewed-on: http://review.membase.org/4244
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>