fixed analyze_core behavior on newer Linux kernels
This was started as removal of ksh dependency and this patch removes
that annoying dependency as a side effect. This is last script that
made our rpms depend of /bin/ksh.

The problem with newer kernels is that they dump cores on which 'file
core' prints whole cmdline, not just executable name. But full
executable name is dumped as part of .auxv section. Unfortunately on
older kernels .auxv doesn't have executable name so we try newer
approach and fall back to older if it doesn't work.

Tested on Debian Sid x86/amd64 (kernel version 2.6.36), RHEL 6 x86_64
(kernel version 2.6.32) and CentOS 5.5 i686 (kernel version 2.6.18).

Note that this utility doesn't work on CentOS 5.2 because it's file
utility is broken w.r.t core dumps. Previous version of this code has
same issue.

Change-Id: I588ce240dcb9ad3278833068f08e02c5f1b9791d
Reviewed-on: http://review.membase.org/4268
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Steve Yen <steve.yen@gmail.com>