MB-3317 - Do not block tap bg fetches during the disk backfill phase.
A new item can be arrived, persisted, and then ejected from memory
while the disk backfill is running. As all tap background fetches
for a given vbucket are skipped during the disk backfill for that
vbucket, there is a possible case where a new item evicted from
memory is not identified by the disk backfill.

This fixes the above issue by (1) allowing tap background fetches
during the disk backfill phase and (2) skipping all the non-resident
items identfied during the memory backfill, which will be fetched
by the disk backfill.

Change-Id: I949294213a740ee9d0ebeea00682a5e5e742bb70
Reviewed-on: http://review.membase.org/4269
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Reviewed-by: Dustin Sallings <dustin@spy.net>