Clean up post-win32 port.
We don't need to conditionally include memcached/engine.h.  We need it
or we don't.

We definitely do not need to copy a whole logger into the file (and
most definintely not below main()) when there's already one ready to
link in.  I did 680e81f specifically to avoid cutting and pasting a
logger in here.

Change-Id: Iaaf3418ffe4608803af555ae07f9f2c99d1cb899
Reviewed-on: http://review.membase.org/4290
Reviewed-by: Bin Cui <bin.cui@gmail.com>
Tested-by: Bin Cui <bin.cui@gmail.com>