MB-3355 - Handle migration from older schemas.
This modifies the source DB by adding the new columns.  It does this
unconditionally and swallows the errors if it fails (usually because
they existed).  If the modifications failed for some other reason,
we'd be left with databases we can't convert, and already log that.

According to SQLite, this is an O(1) operation:

http://www.sqlite.org/lang_altertable.html

    The execution time of the ALTER TABLE command is independent of
    the amount of data in the table. The ALTER TABLE command runs as
    quickly on a table with 10 million rows as it does on a table
    with 1 row.

Change-Id: I660b0419c70b83d1c0cc0791be17407654ebb6f6
Reviewed-on: http://review.membase.org/4302
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>