Allow response messages containing null bytes
We're currently calling strlen() to determine the length of the body
for the response message of the ep-engine custom commands (handled by
the function  EvpUnknownCommand). For the upcoming SYNC command, there's
a need to include null bytes inside the the response message body, so
strlen() truncates the message.

The following change allows each custom command function to explicitly set
the message body size. If not set, it fallbacks to a call to strlen() to
determine the message size.

Change-Id: I88a03d43dbfe8cf6279a3a9b90f21ea716e96d4f
Reviewed-on: http://review.membase.org/4594
Tested-by: Dustin Sallings <dustin@spy.net>
Reviewed-by: Dustin Sallings <dustin@spy.net>