Don't honor reqs_per_tap_event during purging of input messages
The tap log is sent in chunks of messages to improve the network
throughput (sending full packets instead of small packet fragments).
If the consumer starts nack'ing all of the messages we'll end up
in a situation where we don't process all of the nack messages since
we allow the server to send up to 10 times as many requests as we
process input messages before backing off.

We got a finite number of input messages, so we should process
all available input messages if we're backing off.