Reduce the default number of tap events to process
The default number of tap messages to process before switching to the
next tap connection used to be 10k of notifications. If you connect
a high number of tap connections to the server with a large amount of
data to be send (and all of the memory is available in memory) you would
end up in a situation where we wouldn't pay much attention to the input
queue.