Prevent race conditions on SYNC on persistence
Make sure we register the SyncListener instance with the SyncRegistry
before checking if all the keys are dirty or not. If a key was detected
as dirty, but the SyncRegistry registration happened after the key got
persisted to disk, we would have missed that event and the client would
be blocked until the key got persisted again (if ever).

Change-Id: I3db8f90edc462bf28fb12519a555b020977cf31a
Reviewed-on: http://review.membase.org/4764
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>