MB-3489: Track total cache stats in the hash table.
This removes a global stat in favor of a hashtable-specific stat.  It
was previously impossible to properly track the amount of non-resident
data being held, so we instead reduced the stats by the total count.
Since that's not the same as how we incremented it, it was wrong.

Logically, as many stats as can be moved to be hashtable-specific
should be so stats stay updated as vbuckets are added and removed.

Change-Id: I16bf0c55a904199d13cdb9d5d9c9ed8d8706a999
Reviewed-on: http://review.membase.org/4799
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Trond Norbye <trond.norbye@gmail.com>