Set the current CAS for each keyspec in a SYNC response
Currently for the SYNC command, on some cases, for each keyspec in the
response sent to the client, we set the CAS to the same value that the
user specified in the request message (SYNC on persistence) while for
others we set the CAS of the key when the SyncRegistry was notified.
With this patch we get a more consistent behaviour, besides being
eventually more useful to clients.

Change-Id: If0a93bb2aa0af157b248581c4718d5de82231c3e
Reviewed-on: http://review.membase.org/4855
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>