Allow a slave node to create a new checkpoint in some cases.
A slave node should be able to create a new checkpoint based
on the number of items or time bound, in the following cases:

1) A slave node connects to the master server without passing
"TAP_CONNECT_CHECKPOINT connection flag, which means that the slave
doesn't need to synchronize checkpoints with the server.
2) If we need to migrate the data from the 1.6.x master to 1.7.x
slave, the slave will not receive any checkpoint-related messages.

Change-Id: I96b445a17fe78810460ecfc77b9a94322449467c
Reviewed-on: http://review.membase.org/5035
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Trond Norbye <trond.norbye@gmail.com>