fixed bug found by GCC -Werror
GCC suggested to parenthesize && and failed because of -Werror. Closer
inspection revealed that we _do_ want to return not-my-vbucket for
missing vbuckets otherwise we'll face SIGSEGV of death. So we want to
put parenthesis around ||, but I decided to simplify expression by
applying De Morgan's law and 'pushing down' negation through
conditions.

Change-Id: Idce75b7cabe25e20b613802fd06eb9c4f6540c41
Reviewed-on: http://review.membase.org/5196
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>