Incremental backup client stops at the open checkpoint.
Incremental backup (one of the registerd TAP clients) does not
need to receive live mutations from the open checkpoint because
it might receive more than one mutation for the same key from
the open checkpoint.

This change supports an additional option for the registered TAP
client so that it can specify whether it wants to receive mutations
from the open checkpoint or not.

Change-Id: Id7f0b642807bb7f69cb44775437d1b258bcc4a73
Reviewed-on: http://review.membase.org/5209
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>