MB-3568 Send checkpoint_end message after all items are acked.
Checkpoint synchronization between the master and the slave does
not work correctly in the following scenario:
1. The last few items for a given checkpoint were nacked from the
slave, but the master sends the checkpoint_end message to the slave.
2. The master resends those nacked items to the slave, but
the slave already closed the checkpoint and rejected those items.

To resolve this issue, the master should explicitly request an ACK
for the last item as well as any items nacked after the TAP cursor
reached to the checkpoint end.

Change-Id: I693066b963d6716894a61ed154311a1f5d24c6a3
Reviewed-on: http://review.membase.org/5290
Tested-by: Bin Cui <bin.cui@gmail.com>
Reviewed-by: Bin Cui <bin.cui@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>