MB-3603 Fix to the checkpoint synchronization failure.
TAP sequence number is uint32_t type and is reset to 0 when it
reaches to the maximum value. When the sequence number is reset to 0,
we should set the sequence number to 1 and reset the last sequence
number for each vbucket to 0. In this way, when we receive
a subsequent ack, we can determine if a given checkpoint is
synchronized, by comparing the current acked sequence number with the
last sequence number for the corresponding vbucket.

Change-Id: I922d5698f291f44a24d62ef7dfeebb02b44a7c72
Reviewed-on: http://review.membase.org/5406
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>